requirements.txt
clean-text==0.6.0
instagrapi==1.16.30
pandas==1.5.2
PyDrive==1.3.1
photoshop-python-api==0.20.1
python-Levenshtein==0.20.9
thefuzz==0.19.0
termcolor==2.1.1
rich==13.3.1
art==5.8
